requirements.txt
graphrag
openai>=1.0.0
python-dotenv>=1.0.0
rdflib>=7.0.0
numpy>=1.24.0
torch>=2.0.0
pykeen>=1.10.0
